thynkpro-api/requirements.txt
google-generativeai
Flask
dotenv
python-dotenv
Flask-SQLAlchemy
psycopg2
Flask-Migrate
